made pip optional

+++ update.sh
@@ -3,8 +3,10 @@
 pushd $1
 
 source ../venv/bin/activate
-pip freeze | xargs pip uninstall -y
-pip install -r requirements.txt
+if [ "$2" == "pip" ]; then
+    pip freeze | xargs pip uninstall -y
+    pip install -r requirements.txt
+fi
 python manage.py migrate --noinput
 zappa update
 deactivate
